# native/internal/openapi/swagger.yaml
basePath: /
definitions:
...
  github_com_gcc798_microservice-kit_application_iam_internal_response.Response:
    description: API 统一响应格式
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 响应消息
        example: 操作成功
        type: string
    type: object
...
  request.BatchDeleteOrgsRequest:
    properties:
      ids:
        description: 组织ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  request.CreateOrgRequest:
    properties:
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        maxLength: 50
        type: string
      orgCode:
        description: 组织编码
        maxLength: 30
        minLength: 2
        type: string
      orgName:
        description: 组织名称
        maxLength: 50
        minLength: 2
        type: string
      orgType:
        description: 组织类型
        enum:
        - company
        - department
        - group
        type: string
      parentId:
        description: 父组织ID，0表示根组织
        type: integer
      phone:
        description: 联系电话
        maxLength: 11
        minLength: 11
        type: string
      remark:
        description: 备注
        maxLength: 500
        type: string
      sort:
        description: 显示顺序
        type: integer
      status:
        description: 状态：0正常 1停用
        enum:
        - 0
        - 1
        type: integer
    required:
    - orgCode
    - orgName
    type: object
...
  request.PageOrgsRequest:
    properties:
      isAsc:
        description: 'IsAsc 排序方向，asc或desc，支持多个，逗号分隔，如: "asc,desc"'
        type: string
      orderByColumn:
        description: 'OrderByColumn 排序列，支持多列排序，逗号分隔，如: "id,createTime"'
        type: string
      orgCode:
        description: 组织编码
        type: string
      orgName:
        description: 组织名称（模糊查询）
        type: string
      pageNum:
        description: PageNum 当前页码，从1开始
        type: integer
      pageSize:
        description: PageSize 每页数量
        type: integer
      parentId:
        description: 父组织ID（可选）
        type: integer
      status:
        description: 状态：0正常 1停用
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  request.UpdateOrgRequest:
    properties:
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        maxLength: 50
        type: string
      orgCode:
        description: 组织编码
        maxLength: 30
        minLength: 2
        type: string
      orgName:
        description: 组织名称
        maxLength: 50
        minLength: 2
        type: string
      orgType:
        description: 组织类型
        enum:
        - company
        - department
        - group
        type: string
      parentId:
        description: 父组织ID
        type: integer
      phone:
        description: 联系电话
        maxLength: 11
        minLength: 11
        type: string
      remark:
        description: 备注
        maxLength: 500
        type: string
      sort:
        description: 显示顺序
        type: integer
      status:
        description: 状态：0正常 1停用
        enum:
        - 0
        - 1
        type: integer
    type: object
...
host: localhost:9009
info:
  contact: {}
  description: microservice-kit RESTful API
  title: microservice-kit API
  version: "1.0"
paths:
...
  /api/v1/org:
    post:
      consumes:
      - application/json
      description: 创建新组织，需要管理员权限
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建组织请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回组织ID
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      security:
      - Bearer: []
      summary: 创建组织
      tags:
      - 组织管理
  /api/v1/org/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定组织，需要管理员权限
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "404":
          description: 组织不存在
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      security:
      - Bearer: []
      summary: 删除组织
      tags:
      - 组织管理
    get:
      consumes:
      - application/json
      description: 根据组织ID查询组织详细信息
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "404":
          description: 组织不存在
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      security:
      - Bearer: []
      summary: 查询组织详情
      tags:
      - 组织管理
    put:
      consumes:
      - application/json
      description: 更新指定组织的信息，需要管理员权限
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新组织请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "404":
          description: 组织不存在
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      security:
      - Bearer: []
      summary: 更新组织信息
      tags:
      - 组织管理
  /api/v1/org/batch:
    delete:
      consumes:
      - application/json
      description: 批量删除多个组织，需要管理员权限
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 批量删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.BatchDeleteOrgsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      security:
      - Bearer: []
      summary: 批量删除组织
      tags:
      - 组织管理
  /api/v1/org/page:
    post:
      consumes:
      - application/json
      description: 使用 Paginator 分页查询组织列表
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.PageOrgsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      security:
      - Bearer: []
      summary: 分页查询组织列表
      tags:
      - 组织管理
  /api/v1/org/tree:
    get:
      consumes:
      - application/json
      description: 获取完整的组织树结构
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_gcc798_microservice-kit_application_iam_internal_response.Response'
      security:
      - Bearer: []
      summary: 获取组织树
      tags:
      - 组织管理
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"